Move HarmonicDamping to its own file.
