+++ physics/geopotential.hpp
@@ -8,6 +8,7 @@
 #include "geometry/named_quantities.hpp"
 #include "numerics/polynomial.hpp"
 #include "numerics/polynomial_evaluators.hpp"
+#include "physics/harmonic_damping.hpp"
 #include "physics/oblate_body.hpp"
 #include "quantities/named_quantities.hpp"
 
@@ -30,47 +31,6 @@ using quantities::Inverse;
 using quantities::Length;
 using quantities::Quotient;
 using quantities::Square;
-
-// Specification of the damping of a spherical harmonic, acting as a radial
-// multiplier on the potential:
-//   V_damped = σ(‖r‖) V(r).
-class HarmonicDamping final {
- public:
-  HarmonicDamping() = default;
-  explicit HarmonicDamping(Length const& inner_threshold);
-
-  // Above this threshold, the contribution to the potential from this
-  // harmonic is 0, i.e., σ = 0.
-  Length const& outer_threshold() const;
-  // Below this threshold, the contribution to the potential from this
-  // harmonic is undamped, σ = 1.
-  // This class depends on the invariant: outer_threshold = 3 * inner_threshold.
-  Length const& inner_threshold() const;
-
-  // Sets σℜ_over_r and grad_σℜ according to σ as defined by |*this|.
-  template<typename Frame>
-  void ComputeDampedRadialQuantities(
-      Length const& r_norm,
-      Square<Length> const& r²,
-      Vector<double, Frame> const& r_normalized,
-      Inverse<Square<Length>> const& ℜ_over_r,
-      Inverse<Square<Length>> const& ℜʹ,
-      Inverse<Square<Length>>& σℜ_over_r,
-      Vector<Inverse<Square<Length>>, Frame>& grad_σℜ) const;
-
- private:
-  Length outer_threshold_ = Infinity<Length>;
-  Length inner_threshold_ = Infinity<Length>;
-
-  // For r in [outer_threshold, inner_threshold], σ is a polynomial with the
-  // following coefficients in monomial basis.
-  // The constant term is always 0, and is thus ignored in the evaluation.
-  // TODO(phl): We have to specify an evaluator, but we do not use it; we use a
-  // custom evaluation that ignores the constant term instead.  See #1922.
-  PolynomialInMonomialBasis<
-      double, Length, 3,
-      numerics::EstrinEvaluator>::Coefficients sigmoid_coefficients_;
-};
 
 // Representation of the geopotential model of an oblate body.
 template<typename Frame>
