+++ physics/geopotential_test.cpp
@@ -383,49 +383,6 @@ TEST_F(GeopotentialTest, CppF90Comparison) {
   }
 }
 
-TEST_F(GeopotentialTest, HarmonicDamping) {
-  HarmonicDamping σ(1 * Metre);
-  EXPECT_THAT(σ.inner_threshold(), Eq(1 * Metre));
-  EXPECT_THAT(σ.outer_threshold(), Eq(3 * Metre));
-  Vector<double, World> x({1, 0, 0});
-  Inverse<Square<Length>> const ℜ_over_r = 5 / Pow<2>(Metre);
-  Inverse<Square<Length>> const ℜʹ = 17 / Pow<2>(Metre);
-  Inverse<Square<Length>> σℜ_over_r;
-  Vector<Inverse<Square<Length>>, World> grad_σℜ;
-
-  {
-    Length const r = 3 * Metre;
-    σ.ComputeDampedRadialQuantities(
-        r, r * r, x, ℜ_over_r, ℜʹ, σℜ_over_r, grad_σℜ);
-    EXPECT_THAT(σℜ_over_r, Eq(0 / Pow<2>(Metre)));
-    EXPECT_THAT(grad_σℜ.coordinates().x, Eq(0 / Pow<2>(Metre)));
-  }
-  {
-    Length const r = 2 * Metre;
-    σ.ComputeDampedRadialQuantities(
-        r, r * r, x, ℜ_over_r, ℜʹ, σℜ_over_r, grad_σℜ);
-    auto const ℜ = ℜ_over_r * r;
-    auto const σ = 0.5;
-    auto const σʹ = -3 / (4 * Metre);
-    EXPECT_THAT(σℜ_over_r, Eq(ℜ_over_r / 2));
-    EXPECT_THAT(grad_σℜ.coordinates().x, Eq(σʹ * ℜ + ℜʹ * σ));
-  }
-  {
-    Length const r = 1 * Metre;
-    σ.ComputeDampedRadialQuantities(
-        r, r * r, x, ℜ_over_r, ℜʹ, σℜ_over_r, grad_σℜ);
-    EXPECT_THAT(σℜ_over_r, Eq(ℜ_over_r));
-    EXPECT_THAT(grad_σℜ.coordinates().x, Eq(ℜʹ));
-  }
-  {
-    Length const r = 0.5 * Metre;
-    σ.ComputeDampedRadialQuantities(
-        r, r * r, x, ℜ_over_r, ℜʹ, σℜ_over_r, grad_σℜ);
-    EXPECT_THAT(σℜ_over_r, Eq(ℜ_over_r));
-    EXPECT_THAT(grad_σℜ.coordinates().x, Eq(ℜʹ));
-  }
-}
-
 TEST_F(GeopotentialTest, ThresholdComputation) {
   SolarSystem<ICRS> solar_system_2000(
             SOLUTION_DIR / "astronomy" / "sol_gravity_model.proto.txt",
