+++ physics/physics.vcxproj
@@ -45,6 +45,8 @@
     <ClInclude Include="discrete_trajectory_types_body.hpp" />
     <ClInclude Include="equipotential.hpp" />
     <ClInclude Include="equipotential_body.hpp" />
+    <ClInclude Include="harmonic_damping.hpp" />
+    <ClInclude Include="harmonic_damping_body.hpp" />
     <ClInclude Include="mechanical_system.hpp" />
     <ClInclude Include="mechanical_system_body.hpp" />
     <ClInclude Include="continuous_trajectory_body.hpp" />
@@ -108,6 +110,7 @@
     <ClCompile Include="discrete_trajectory_segment_test.cpp" />
     <ClCompile Include="discrete_trajectory_test.cpp" />
     <ClCompile Include="equipotential_test.cpp" />
+    <ClCompile Include="harmonic_damping_test.cpp" />
     <ClCompile Include="mechanical_system_test.cpp" />
     <ClCompile Include="continuous_trajectory_test.cpp" />
     <ClCompile Include="degrees_of_freedom_test.cpp" />
