+++ physics/physics.vcxproj.filters
@@ -221,6 +221,12 @@
     <ClInclude Include="equipotential_body.hpp">
       <Filter>Source Files</Filter>
     </ClInclude>
+    <ClInclude Include="harmonic_damping.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="harmonic_damping_body.hpp">
+      <Filter>Source Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="degrees_of_freedom_test.cpp">
@@ -329,6 +335,9 @@
       <Filter>Test Files</Filter>
     </ClCompile>
     <ClCompile Include="equipotential_test.cpp">
+      <Filter>Test Files</Filter>
+    </ClCompile>
+    <ClCompile Include="harmonic_damping_test.cpp">
       <Filter>Test Files</Filter>
     </ClCompile>
   </ItemGroup>
